+++ packages/rice-uikit/src/components/Spinner/Spinner.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import styled, { keyframes } from "styled-components";
 import RiceIcon from "./RiceIcon";
-// import PancakeIcon from "./PancakeIcon";
 import { SpinnerProps } from "./types";
 
 const rotate = keyframes`
@@ -13,49 +12,21 @@ const rotate = keyframes`
   }
 `;
 
-// const float = keyframes`
-// 	0% {
-// 		transform: translatey(0px);
-// 	}
-// 	50% {
-// 		transform: translatey(10px);
-// 	}
-// 	100% {
-// 		transform: translatey(0px);
-// 	}
-// `;
-
 const Container = styled.div`
   position: relative;
 `;
 
-// const RotatingPancakeIcon = styled(PancakeIcon)`
-//   position: absolute;
-//   top: 0;
-//   left: 0;
-//   animation: ${rotate} 2s linear infinite;
-//   transform: translate3d(0, 0, 0);
-// `;
-
-// const FloatingPanIcon = styled(PanIcon)`
-//   animation: ${float} 6s ease-in-out infinite;
-//   transform: translate3d(0, 0, 0);
-// `;
-const FloatingPanIconWrapper = styled.div`
-  position: absolute;
+const RotatingRiceIcon = styled(RiceIcon)`
   top: 0;
   left: 0;
   animation: ${rotate} 2s linear infinite;
   transform: translate3d(0, 0, 0);
-`;
+`
 
 const Spinner: React.FC<SpinnerProps> = ({ size = 128 }) => {
   return (
     <Container>
-      {/* <RotatingPancakeIcon width={`${size * 0.5}px`} /> */}
-      <FloatingPanIconWrapper>
-        <RiceIcon width={`${size}px`} />
-      </FloatingPanIconWrapper>
+      <RotatingRiceIcon width={`${size}px`} />
     </Container>
   );
 };