+++ packages/rice-uikit/src/components/SpinnerCake/index.tsx
@@ -1,0 +1,2 @@
+export { default as SpinnerCake } from "./SpinnerCake";
+export type { SpinnerCakeProps } from "./types";